+++ packages/dso-toolkit/src/components/mark-bar/mark-bar.args.ts
@@ -5,18 +5,21 @@ import { HandlerFunction } from "@storybook/addon-actions/*";
 
 export interface MarkBarArgs {
   value: string;
-  label?: string;
+  label: string;
   current: number;
   totalCount: number;
+  focus: boolean;
   dsoInput: HandlerFunction;
   dsoNext: HandlerFunction;
   dsoPrevious: HandlerFunction;
   dsoClear: HandlerFunction;
 }
 
-export const markBarArgs: Omit<MarkBarArgs, "value" | "label" | "dsoInput" | "dsoNext" | "dsoPrevious" | "dsoClear"> = {
+export const markBarArgs: Omit<MarkBarArgs, "value" | "dsoInput" | "dsoNext" | "dsoPrevious" | "dsoClear"> = {
+  label: "Zoeken binnen gehele document, en verder dan dat.",
   current: 1,
   totalCount: 8,
+  focus: false,
 };
 
 export const markBarArgTypes: ArgTypes<MarkBarArgs> = {
@@ -44,9 +47,22 @@ export const markBarArgTypes: ArgTypes<MarkBarArgs> = {
   dsoClear: {
     action: "dsoClear",
   },
+  focus: {
+    type: "boolean",
+  },
 };
 
 export function markBarArgsMapper(a: MarkBarArgs): MarkBar {
+  if (a.focus) {
+    setTimeout(() => {
+      const markBar = document.querySelector<HTMLElement & { dsoFocus: (options: { select: boolean }) => void }>(
+        "dso-mark-bar",
+      );
+
+      markBar?.dsoFocus({ select: true });
+    });
+  }
+
   return {
     ...a,
   };